--- a/deltametrics/_version.py
+++ b/deltametrics/_version.py
@@ -2,4 +2,4 @@ def __version__():
     """
     Private version declaration, gets assigned to dm.__version__ during import
     """
-    return "0.4.2"
+    return "0.4.3"